+++ crates/voicevox_core/src/engine/acoustic_feature_extractor.rs
@@ -196,8 +196,11 @@ pub enum Phoneme {
     ConsonantZ,
 }
 
-#[derive(Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, derive_more::Display)]
-pub(crate) enum Consonant {
+/// 子音。
+#[derive(
+    Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, derive_more::Display, SerializeDisplay,
+)]
+pub enum Consonant {
     /// `b`。
     #[display("b")]
     B,
@@ -327,8 +330,11 @@ pub(crate) enum Consonant {
     Z,
 }
 
-#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, derive_more::Display)]
-pub(crate) enum NonConsonant {
+/// 子音ではない音素。
+#[derive(
+    Clone, PartialEq, Eq, PartialOrd, Ord, Hash, Debug, derive_more::Display, SerializeDisplay,
+)]
+pub enum NonConsonant {
     /// `pau`。
     #[display("pau")]
     MorablePau,