+++ crates/voicevox_core/src/error.rs
@@ -268,8 +268,8 @@ pub(crate) enum InvalidQueryErrorSource {
     #[error("\"sil\"を含む文字列である必要があります")]
     MustContainSil,
 
-    #[error("`0`にすることはできません")]
-    IsZero,
+    #[error("アクセント位置はモーラ数を超過してはいけません")]
+    TooLargeAccent,
 
     #[error("0より大きい{DEFAULT_SAMPLING_RATE}の倍数でなければなりません")]
     IsNotMultipleOfBaseSamplingRate,