+++ crates/voicevox_core/src/synthesizer.rs
@@ -13,7 +13,7 @@ use std::{
     sync::Arc,
 };
 use tracing::info;
-use typed_floats::{NonNaNFinite, PositiveFinite};
+use typed_floats::{NonNaNFinite, PositiveFinite, tf32};
 
 use crate::{
     AccentPhrase, AudioQuery, OnExistingVoiceModelId, Result, StyleId, VoiceModelId,
@@ -57,7 +57,7 @@ use crate::{
     },
     error::ErrorRepr,
     future::FutureExt as _,
-    numerics::positive_finite_f32,
+    numerics::{non_nan_finite_f32, positive_finite_f32},
 };
 
 pub const DEFAULT_CPU_NUM_THREADS: u16 = 0;
@@ -523,7 +523,7 @@ trait AsInner {
         let pcm = self.render(&audio, 0..audio.frame_length).await?;
         Ok(wav_from_s16le(
             &pcm,
-            audio_query.output_sampling_rate,
+            audio_query.output_sampling_rate.get().get(),
             audio_query.output_stereo,
         ))
     }
@@ -607,7 +607,7 @@ trait AsInner {
 
         #[ext]
         impl<P: Clone> LengthedPhoneme<P> {
-            fn with_length(&self, length: f32) -> LengthedPhoneme<P> {
+            fn with_length(&self, length: PositiveFinite<f32>) -> LengthedPhoneme<P> {
                 LengthedPhoneme {
                     phoneme: self.phoneme.clone(),
                     length,
@@ -679,7 +679,7 @@ trait AsInner {
 
         for i in 0..vowel_phoneme_data_list.len() {
             if vowel_phoneme_data_list[i].is_unvoiced() {
-                f0_list[i] = 0.;
+                f0_list[i] = tf32::ZERO.into();
             }
         }
 
@@ -1042,7 +1042,7 @@ trait AsInner {
         &self,
         phoneme_vector: &[i64],
         style_id: StyleId,
-    ) -> Result<Vec<f32>> {
+    ) -> Result<Vec<PositiveFinite<f32>>> {
         let status = self.status().clone();
         let phoneme_vector = ndarray::arr1(phoneme_vector);
         status
@@ -1065,7 +1065,7 @@ trait AsInner {
         start_accent_phrase_vector: &[i64],
         end_accent_phrase_vector: &[i64],
         style_id: StyleId,
-    ) -> Result<Vec<f32>> {
+    ) -> Result<Vec<NonNaNFinite<f32>>> {
         let status = self.status().clone();
         let vowel_phoneme_vector = ndarray::arr1(vowel_phoneme_vector);
         let consonant_phoneme_vector = ndarray::arr1(consonant_phoneme_vector);
@@ -1184,7 +1184,7 @@ impl<R: InferenceRuntime> Status<R> {
         &self,
         phoneme_vector: ndarray::Array1<i64>,
         style_id: StyleId,
-    ) -> Result<Vec<f32>> {
+    ) -> Result<Vec<PositiveFinite<f32>>> {
         // `TalkDomain`と`ExperimentalTalkDomain`の両方がある場合、`TalkDomain`を優先
         if self.contains_domain::<TalkDomain>(style_id) {
             let (model_id, inner_voice_id) = self.ids_for::<TalkDomain>(style_id)?;
@@ -1200,7 +1200,10 @@ impl<R: InferenceRuntime> Status<R> {
                     A::LIGHT_INFERENCE_CANCELLABLE,
                 )
                 .await?;
-            return Ok(ensure_minimum_phoneme_length(output.into_vec()));
+            let output = ensure_non_nan_finite(&output.into_vec(), |invalid| {
+                anyhow!("`predict_duration` returned an array that contains: {invalid}")
+            })?;
+            return Ok(ensure_minimum_phoneme_length(output));
         }
         let (model_id, inner_voice_id) = self.ids_for::<ExperimentalTalkDomain>(style_id)?;
 
@@ -1216,7 +1219,10 @@ impl<R: InferenceRuntime> Status<R> {
                 A::LIGHT_INFERENCE_CANCELLABLE,
             )
             .await?;
-        Ok(ensure_minimum_phoneme_length(output.into_vec()))
+        let output = ensure_non_nan_finite(&output.into_vec(), |invalid| {
+            anyhow!("`predict_duration` returned an array that contains: {invalid}")
+        })?;
+        Ok(ensure_minimum_phoneme_length(output))
     }
 
     #[expect(
@@ -1234,7 +1240,7 @@ impl<R: InferenceRuntime> Status<R> {
         start_accent_phrase_vector: ndarray::Array1<i64>,
         end_accent_phrase_vector: ndarray::Array1<i64>,
         style_id: StyleId,
-    ) -> Result<Vec<f32>> {
+    ) -> Result<Vec<NonNaNFinite<f32>>> {
         // `TalkDomain`と`ExperimentalTalkDomain`の両方がある場合、`TalkDomain`を優先
         if self.contains_domain::<TalkDomain>(style_id) {
             let (model_id, inner_voice_id) = self.ids_for::<TalkDomain>(style_id)?;
@@ -1254,7 +1260,9 @@ impl<R: InferenceRuntime> Status<R> {
                     A::LIGHT_INFERENCE_CANCELLABLE,
                 )
                 .await?;
-            return Ok(output.into_vec());
+            return ensure_non_nan_finite(&output.into_vec(), |invalid| {
+                anyhow!("`predict_intonation` returned an array that contains: {invalid}")
+            });
         }
         let (model_id, inner_voice_id) = self.ids_for::<ExperimentalTalkDomain>(style_id)?;
 
@@ -1275,7 +1283,9 @@ impl<R: InferenceRuntime> Status<R> {
             )
             .await?;
 
-        Ok(output.into_vec())
+        ensure_non_nan_finite(&output.into_vec(), |invalid| {
+            anyhow!("`predict_intonation` returned an array that contains: {invalid}")
+        })
     }
 
     /// モデル`generate_full_intermediate`の実行と、その前後の処理を行う。
@@ -1583,13 +1593,13 @@ impl Default for AudioQuery {
     fn default() -> Self {
         Self {
             accent_phrases: vec![],
-            speed_scale: 1.,
-            pitch_scale: 0.,
-            intonation_scale: 1.,
-            volume_scale: 1.,
-            pre_phoneme_length: 0.1,
-            post_phoneme_length: 0.1,
-            output_sampling_rate: DEFAULT_SAMPLING_RATE,
+            speed_scale: positive_finite_f32!(1.),
+            pitch_scale: non_nan_finite_f32!(0.),
+            intonation_scale: non_nan_finite_f32!(1.),
+            volume_scale: positive_finite_f32!(1.),
+            pre_phoneme_length: positive_finite_f32!(0.1),
+            post_phoneme_length: positive_finite_f32!(0.1),
+            output_sampling_rate: Default::default(),
             output_stereo: false,
             kana: None,
         }
@@ -2285,7 +2295,11 @@ pub(crate) mod blocking {
             phoneme_vector: &[i64],
             style_id: StyleId,
         ) -> crate::Result<Vec<f32>> {
-            self.0.predict_duration(phoneme_vector, style_id).block_on()
+            // TODO: typed_floatsにissueかPRを出しに行き、スライス変換かbytemuck対応を入れてもらう
+            self.0
+                .predict_duration(phoneme_vector, style_id)
+                .block_on()
+                .map(|o| o.into_iter().map(Into::into).collect())
         }
 
         pub fn predict_intonation(
@@ -2299,6 +2313,7 @@ pub(crate) mod blocking {
             end_accent_phrase_vector: &[i64],
             style_id: StyleId,
         ) -> crate::Result<Vec<f32>> {
+            // TODO: typed_floatsにissueかPRを出しに行き、スライス変換かbytemuck対応を入れてもらう
             self.0
                 .predict_intonation(
                     length,
@@ -2311,6 +2326,7 @@ pub(crate) mod blocking {
                     style_id,
                 )
                 .block_on()
+                .map(|o| o.into_iter().map(Into::into).collect())
         }
 
         pub fn generate_full_intermediate(
@@ -3415,16 +3431,18 @@ pub(crate) mod nonblocking {
 
 #[cfg(test)]
 mod tests {
-    use std::mem;
+    use std::{mem, num::NonZero};
 
     use super::{AccelerationMode, AsInner as _, DEFAULT_HEAVY_INFERENCE_CANCELLABLE};
     use crate::{
         AccentPhrase, FramePhoneme, Note, NoteId, Result, Score, StyleId,
         asyncs::BlockingThreadPool, engine::talk::Mora, macros::tests::assert_debug_fmt_eq,
+        numerics::non_zero,
     };
     use ::test_util::OPEN_JTALK_DIC_DIR;
     use itertools::Itertools as _;
     use rstest::rstest;
+    use typed_floats::tf32;
 
     #[rstest]
     #[case(Ok(()))]
@@ -3761,12 +3779,17 @@ mod tests {
         assert_eq!(result.unwrap().len(), F0_LENGTH * 256);
     }
 
-    type TextConsonantVowelData =
-        [(&'static [(&'static str, &'static str, &'static str)], usize)];
+    type TextConsonantVowelData = [(
+        &'static [(&'static str, &'static str, &'static str)],
+        NonZero<usize>,
+    )];
 
     // [([(テキスト, 母音, 子音), ...], アクセントの位置), ...] の形式
     const TEXT_CONSONANT_VOWEL_DATA1: &TextConsonantVowelData = &[
-        (&[("コ", "k", "o"), ("レ", "r", "e"), ("ワ", "w", "a")], 3),
+        (
+            &[("コ", "k", "o"), ("レ", "r", "e"), ("ワ", "w", "a")],
+            non_zero!(3usize),
+        ),
         (
             &[
                 ("テ", "t", "e"),
@@ -3775,12 +3798,15 @@ mod tests {
                 ("デ", "d", "e"),
                 ("ス", "s", "U"),
             ],
-            1,
+            non_zero!(1usize),
         ),
     ];
 
     const TEXT_CONSONANT_VOWEL_DATA2: &TextConsonantVowelData = &[
-        (&[("コ", "k", "o"), ("レ", "r", "e"), ("ワ", "w", "a")], 1),
+        (
+            &[("コ", "k", "o"), ("レ", "r", "e"), ("ワ", "w", "a")],
+            non_zero!(1usize),
+        ),
         (
             &[
                 ("テ", "t", "e"),
@@ -3789,7 +3815,7 @@ mod tests {
                 ("デ", "d", "e"),
                 ("ス", "s", "U"),
             ],
-            3,
+            non_zero!(3usize),
         ),
     ];
 
@@ -3857,11 +3883,11 @@ mod tests {
                 // NOTE: 子音の長さが必ず非ゼロになるテストケースを想定している
                 assert_ne!(
                     mora.consonant_length,
-                    Some(0.),
+                    Some(tf32::ZERO),
                     "expected mora.consonant_length is not Some(0.0), but got Some(0.0)."
                 );
-                assert_eq!(mora.consonant, Some(consonant.to_string()));
-                assert_eq!(mora.vowel, vowel);
+                assert_eq!(mora.consonant, Some(consonant.parse().unwrap()));
+                assert_eq!(mora.vowel, vowel.parse().unwrap());
                 // NOTE: 母音の長さが必ず非ゼロになるテストケースを想定している
                 assert_ne!(
                     mora.vowel_length, 0.,
@@ -3932,11 +3958,11 @@ mod tests {
                 // NOTE: 子音の長さが必ず非ゼロになるテストケースを想定している
                 assert_ne!(
                     mora.consonant_length,
-                    Some(0.),
+                    Some(tf32::ZERO),
                     "expected mora.consonant_length is not Some(0.0), but got Some(0.0)."
                 );
-                assert_eq!(mora.consonant, Some(consonant.to_string()));
-                assert_eq!(mora.vowel, vowel);
+                assert_eq!(mora.consonant, Some(consonant.parse().unwrap()));
+                assert_eq!(mora.vowel, vowel.parse().unwrap());
                 // NOTE: 母音の長さが必ず非ゼロになるテストケースを想定している
                 assert_ne!(
                     mora.vowel_length, 0.,
@@ -4001,7 +4027,7 @@ mod tests {
             assert_eq!(pause_mora.text, "、");
             assert_eq!(pause_mora.consonant, None);
             assert_eq!(pause_mora.consonant_length, None);
-            assert_eq!(pause_mora.vowel, "pau");
+            assert_eq!(pause_mora.vowel, "pau".parse().unwrap());
             assert_ne!(
                 pause_mora.vowel_length, 0.0,
                 "pause_mora.vowel_length should not be 0.0",